refactor(platform): merge the workdir chip into the sandbox chip
The standalone composer workspace chip read as a duplicate entry for the
Workspace-files panel. The Sandbox pill now doubles as the thread's
sandbox-settings entry: clicking it opens a popover holding the working
directory (extensible with further per-thread sandbox settings later).
A custom workdir surfaces in the pill tooltip; before the thread exists
or while the composer is disabled the pill stays purely informational.
