+++ services/platform/messages/en.json
@@ -1470,8 +1470,6 @@
     },
     "workdir": {
       "label": "Working directory",
-      "chipTooltip": "Where the agent starts each turn. Point it at a repo folder inside the workspace and the agent picks up that repo's own instructions and skills.",
-      "rootLabel": "workspace",
       "placeholder": "e.g. tale",
       "helper": "Relative to the workspace root. Applies from your next message; while the folder doesn't exist yet, the agent runs at the workspace root.",
       "save": "Save",
@@ -1487,6 +1485,8 @@
     "sandbox": {
       "label": "Sandbox",
       "tooltip": "Runs in an isolated sandbox VM with its own persistent workspace — it can run real commands and install packages, safely separated from your computer.",
+      "settingsTitle": "Sandbox settings",
+      "settingsHint": "Click to change this conversation's sandbox settings.",
       "working": "Working",
       "workingTooltip": "The agent is running a task in its sandbox right now.",
       "sleeping": "Sleeping",
